moving webpack into client/ directory. simplifying project structure
