+++ client/components/detail/detail.jsx
