+++ client/routes/routes.jsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import {Route} from 'react-router'
 
-import Dashboard from '../components/dashboard/dashboard'
+import Dashboard from 'components/dashboard/dashboard'
 
 export default function getRoutes() {
   return (
