+++ client/webpack/config.dist.js
@@ -1,7 +1,9 @@
 var webpack = require('webpack')
 var base = require('./config')
-var pkg = require('../package.json')
+var pkg = require('../../package.json')
+
 var ExtractTextPlugin = require('extract-text-webpack-plugin')
+var CompressionPlugin = require('compression-webpack-plugin')
 
 var config = {
   context: base.context,
@@ -9,11 +11,8 @@ var config = {
   output: base.output,
   module: base.module,
   plugins: [
-    new ExtractTextPlugin('[name]-' + pkg.version + '.style.css'),
+    new ExtractTextPlugin('app-' + pkg.version + '.style.css'),
     new webpack.optimize.DedupePlugin(),
-    new webpack.optimize.CommonsChunkPlugin({
-      name: 'common'
-    }),
     new webpack.optimize.UglifyJsPlugin({
       compress: {
         warnings: false
@@ -27,7 +26,8 @@ var config = {
         NODE_ENV: JSON.stringify('production')
       }
     }),
-    new webpack.IgnorePlugin(/^\.\/locale$/, [/moment$/])
+    new webpack.IgnorePlugin(/^\.\/locale$/, [/moment$/]),
+    new CompressionPlugin()
   ],
   resolve: base.resolve
 }
