+++ package.json
@@ -12,16 +12,17 @@
     "dbLocation": "mongodb://localhost/didactic-disco"
   },
   "scripts": {
+    "clean": "rimraf build",
     "lint": "npm run lint:client && npm run lint:server",
     "lint:client": "eslint --ext .jsx,.js client",
     "lint:server": "eslint --ext .js,.json server",
     "lint:fix": "npm run lint -- --fix",
     "pretest": "npm run lint",
-    "test": "echo 'no tests available' && exit 1",
-    "bundle": "NODE_ENV=production webpack --config webpack/config.dist.js --progress",
+    "test": "echo 'no tests available' && exit 0",
+    "bundle": "npm run clean && npm run test && NODE_ENV=production webpack --config client/webpack/config.dist.js --progress",
     "server": "NODE_ENV=production forever ./server/server.js",
     "express-dev-server": "nodemon ./server/server.js",
-    "webpack-dev-server": "webpack-dev-server --colors --progress --hot --inline --config  ./webpack/config.dev.js --port 2992",
+    "webpack-dev-server": "webpack-dev-server --colors --progress --hot --inline --config  client/webpack/config.dev.js --port 2992",
     "start": "NODE_ENV=development parallelshell 'npm run express-dev-server' 'npm run webpack-dev-server'"
   },
   "dependencies": {
@@ -54,6 +55,7 @@
     "babel-preset-es2015": "^6.6.0",
     "babel-preset-react": "^6.5.0",
     "babel-preset-stage-0": "^6.5.0",
+    "compression-webpack-plugin": "^0.3.1",
     "css-loader": "^0.23.1",
     "eslint": "^2.9.0",
     "eslint-plugin-react": "^5.0.1",
@@ -62,6 +64,7 @@
     "nodemon": "^1.9.2",
     "parallelshell": "^2.0.0",
     "postcss-loader": "^0.9.1",
+    "rimraf": "^2.5.2",
     "sass-loader": "^3.2.0",
     "style-loader": "^0.13.1",
     "webpack": "^1.13.0",
