+++ server/views/layout.pug
@@ -1,0 +1,14 @@
+doctype html
+html.no-js(lang='en')
+  head
+    meta(charset='utf-8')
+    meta(name='viewport', content='width=device-width')
+
+    block header
+    block links
+
+  body
+    #app
+
+    block scripts
+
